Bug Fixes
Added if statements so that the new week would not start until Monday. Added if statements to catch instances where the day/week would go to -1 instead of 3 (the end or the array). Added a block on the end to send myself a complete schedule for the week for records every Monday